requirements.txt.txt
Django>=3.1.6   
djangorestframework>=3.12.2
psycopg2>=2.7.5 
Pillow >= 8.0
flake8>=3.8.4